Reworked Damage System
In a none breaking way, I removed the Damage.cs script after moving all of its unique functionality into the Health.cs script.

I like the design pattern of having Damage separate from Health, but at multiple points we ended up just not using it. It literally does nothing right now and it would be more work to change everything that isn't using it than it is to just remove it.