#15 Fixed Critical Health Indicator
I added flag called isCriticalIndicatorVisible to keep tabs on whether the critical indicator is currently being shown. This helps stop us from making extra copies of the indicator when we don't need them.

I tweaked the ShowCriticalIndicator and HideCriticalIndicator parts to use the isCriticalIndicatorVisible flag. This makes sure we only show or hide the indicator when we really have to.